+++ source/projects/py/py.h
@@ -146,4 +146,14 @@ void py_edclose(t_py* x, char** text, long size);
 t_max_err py_edsave(t_py* x, char** text, long size);
 // void py_okclose(t_py* x, char *s, short *result);
 
+/*--------------------------------------------------------------------------*/
+/* max datastructure support methods */
+
+// table
+bool py_table_exists(t_py* x, char* table_name);
+t_max_err py_list_to_table(t_py* x, char* table_name, PyObject* plist);
+PyObject* py_table_to_list(t_py* x, char* table_name);
+
+
+
 #endif // PY_H